- fix for SIS-19 QTreeWriter and GeoRSSData.save don't check to make sure the directory exists before creating data there
git-svn-id: https://svn.apache.org/repos/asf/incubator/sis/trunk@1032858 13f79535-47bb-0310-9956-ffa450edef68

+++ sis-core/src/main/java/org/apache/sis/storage/GeoRSSData.java
@@ -20,6 +20,7 @@ package org.apache.sis.storage;
 //JDK imports
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
+import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.FileWriter;
@@ -105,6 +106,7 @@ public class GeoRSSData implements QuadTreeData {
 	 */
 	public void saveToFile(Item item, GeoRSSModule geoRSSModule,
 			String directory) {
+	  if(!new File(directory).exists()) new File(directory).mkdirs();
 		try {
 			BufferedWriter writer = new BufferedWriter(new FileWriter(directory
 					+ filename));
